// Modules/Angular/Day9-10/Disha-Shah/Assignment/assignmentApp/src/app/models/IStudent.ts
export interface IStudent{
    name: string;
    dob: Date;
    placeOfBirth: string;
    language: string;
    address: string;
    fatherName: string;
    fatherEmail: string;
    fatherQualification: string;
    fatherProfession: string;
    fatherDesignation: string;
    fatherPhone: string;
    motherName: string;
    motherEmail: string;
    motherQualification: string;
    motherProfession: string;
    motherDesignation: string;
    motherPhone: string;
    emergencyContact: Array<any>;
    referenceDetails: Array<any>;
}